remove DML check on getUpdateCount()
add MULTI_STATEMENT_COUNT in test

fix comment

close connection
