+++ src/test/java/net/snowflake/client/jdbc/MultiStatementIT.java
@@ -34,223 +34,6 @@ public class MultiStatementIT extends BaseJDBCTest {
   }
 
   @Test
-  public void testMultiStmtTransaction() throws SQLException {
-    Connection connection = getConnection();
-    Statement statement = connection.createStatement();
-
-    statement.execute(
-        "create or replace table test_multi_txn(c1 number, c2 string)" + " as select 10, 'z'");
-
-    statement.unwrap(SnowflakeStatement.class).setParameter("MULTI_STATEMENT_COUNT", 4);
-    String multiStmtQuery =
-        "begin;\n"
-            + "delete from test_multi_txn;\n"
-            + "insert into test_multi_txn values (1, 'a'), (2, 'b');\n"
-            + "commit";
-
-    boolean hasResultSet = statement.execute(multiStmtQuery);
-    // first statement
-    assertFalse(hasResultSet);
-    assertNull(statement.getResultSet());
-    assertEquals(-1, statement.getUpdateCount());
-
-    // second statement
-    assertFalse(statement.getMoreResults());
-    assertNull(statement.getResultSet());
-    assertEquals(1, statement.getUpdateCount());
-
-    // third statement
-    assertFalse(statement.getMoreResults());
-    assertNull(statement.getResultSet());
-    assertEquals(2, statement.getUpdateCount());
-
-    // fourth statement
-    assertFalse(statement.getMoreResults());
-    assertNull(statement.getResultSet());
-    assertEquals(-1, statement.getUpdateCount());
-
-    assertFalse(statement.getMoreResults());
-    assertEquals(-1, statement.getUpdateCount());
-
-    statement.unwrap(SnowflakeStatement.class).setParameter("MULTI_STATEMENT_COUNT", 1);
-    statement.execute("drop table if exists test_multi_txn");
-    statement.close();
-    connection.close();
-  }
-
-  @Test
-  public void testMultiStmtTransactionRollback() throws SQLException {
-    Connection connection = getConnection();
-    Statement statement = connection.createStatement();
-
-    statement.execute(
-        "create or replace table test_multi_txn_rb(c1 number, c2 string)" + " as select 10, 'z'");
-
-    statement.unwrap(SnowflakeStatement.class).setParameter("MULTI_STATEMENT_COUNT", 4);
-    String multiStmtQuery =
-        "begin;\n"
-            + "delete from test_multi_txn_rb;\n"
-            + "rollback;\n"
-            + "select count(*) from test_multi_txn_rb";
-
-    boolean hasResultSet = statement.execute(multiStmtQuery);
-    // first statement
-    assertFalse(hasResultSet);
-    assertNull(statement.getResultSet());
-    assertEquals(-1, statement.getUpdateCount());
-
-    // second statement
-    assertFalse(statement.getMoreResults());
-    assertNull(statement.getResultSet());
-    assertEquals(1, statement.getUpdateCount());
-
-    // third statement
-    assertFalse(statement.getMoreResults());
-    assertNull(statement.getResultSet());
-    assertEquals(-1, statement.getUpdateCount());
-
-    // fourth statement
-    assertTrue(statement.getMoreResults());
-    assertEquals(-1, statement.getUpdateCount());
-    ResultSet rs = statement.getResultSet();
-    assertTrue(rs.next());
-    assertEquals(1, rs.getInt(1));
-    assertFalse(rs.next());
-
-    assertFalse(statement.getMoreResults());
-    assertEquals(-1, statement.getUpdateCount());
-
-    statement.unwrap(SnowflakeStatement.class).setParameter("MULTI_STATEMENT_COUNT", 1);
-    statement.execute("drop table if exists test_multi_txn_rb");
-    statement.close();
-    connection.close();
-  }
-
-  @Test
-  public void testMultiStmtExecute() throws SQLException {
-    Connection connection = getConnection();
-    Statement statement = connection.createStatement();
-
-    statement.unwrap(SnowflakeStatement.class).setParameter("MULTI_STATEMENT_COUNT", 3);
-    String multiStmtQuery =
-        "create or replace temporary table test_multi (cola int);\n"
-            + "insert into test_multi VALUES (1), (2);\n"
-            + "select cola from test_multi order by cola asc";
-
-    boolean hasResultSet = statement.execute(multiStmtQuery);
-    // first statement
-    assertFalse(hasResultSet);
-    assertNull(statement.getResultSet());
-    assertEquals(-1, statement.getUpdateCount());
-
-    // second statement
-    assertFalse(statement.getMoreResults());
-    assertNull(statement.getResultSet());
-    assertEquals(2, statement.getUpdateCount());
-
-    // third statement
-    assertTrue(statement.getMoreResults());
-    assertEquals(-1, statement.getUpdateCount());
-    ResultSet rs = statement.getResultSet();
-    assertTrue(rs.next());
-    assertEquals(1, rs.getInt(1));
-    assertTrue(rs.next());
-    assertEquals(2, rs.getInt(1));
-    assertFalse(rs.next());
-
-    assertFalse(statement.getMoreResults());
-    assertEquals(-1, statement.getUpdateCount());
-
-    statement.close();
-    connection.close();
-  }
-
-  @Test
-  public void testMultiStmtExecuteUpdate() throws SQLException {
-    Connection connection = getConnection();
-    Statement statement = connection.createStatement();
-    String multiStmtQuery =
-        "create or replace temporary table test_multi (cola int);\n"
-            + "insert into test_multi VALUES (1), (2);\n"
-            + "select cola from test_multi order by cola asc";
-
-    statement.unwrap(SnowflakeStatement.class).setParameter("MULTI_STATEMENT_COUNT", 3);
-    int rowCount = statement.executeUpdate(multiStmtQuery);
-    // first statement
-    assertEquals(0, rowCount);
-    assertNull(statement.getResultSet());
-    assertEquals(-1, statement.getUpdateCount());
-
-    // second statement
-    assertFalse(statement.getMoreResults());
-    assertNull(statement.getResultSet());
-    assertEquals(2, statement.getUpdateCount());
-
-    // third statement
-    assertTrue(statement.getMoreResults());
-    assertEquals(-1, statement.getUpdateCount());
-    ResultSet rs = statement.getResultSet();
-    assertTrue(rs.next());
-    assertEquals(1, rs.getInt(1));
-    assertTrue(rs.next());
-    assertEquals(2, rs.getInt(1));
-    assertFalse(rs.next());
-
-    assertFalse(statement.getMoreResults());
-    assertEquals(-1, statement.getUpdateCount());
-
-    statement.close();
-    connection.close();
-  }
-
-  @Test
-  public void testMultiStmtExecuteQuery() throws SQLException {
-    Connection connection = getConnection();
-    Statement statement = connection.createStatement();
-    String multiStmtQuery =
-        "select 1;\n"
-            + "create or replace temporary table test_multi (cola int);\n"
-            + "insert into test_multi VALUES (1), (2);\n"
-            + "select cola from test_multi order by cola asc";
-
-    statement.unwrap(SnowflakeStatement.class).setParameter("MULTI_STATEMENT_COUNT", 4);
-    ResultSet rs = statement.executeQuery(multiStmtQuery);
-    // first statement
-    assertNotNull(rs);
-    assertNotNull(statement.getResultSet());
-    assertEquals(-1, statement.getUpdateCount());
-    assertTrue(rs.next());
-    assertEquals(1, rs.getInt(1));
-    assertFalse(rs.next());
-
-    // second statement
-    assertFalse(statement.getMoreResults());
-    assertNull(statement.getResultSet());
-    assertEquals(-1, statement.getUpdateCount());
-
-    // third statement
-    assertFalse(statement.getMoreResults());
-    assertNull(statement.getResultSet());
-    assertEquals(2, statement.getUpdateCount());
-
-    // fourth statement
-    assertTrue(statement.getMoreResults());
-    assertEquals(-1, statement.getUpdateCount());
-    rs = statement.getResultSet();
-    assertTrue(rs.next());
-    assertEquals(1, rs.getInt(1));
-    assertTrue(rs.next());
-    assertEquals(2, rs.getInt(1));
-    assertFalse(rs.next());
-
-    assertFalse(statement.getMoreResults());
-    assertEquals(-1, statement.getUpdateCount());
-
-    statement.close();
-    connection.close();
-  }
-
-  @Test
   public void testMultiStmtExecuteUpdateFail() throws SQLException {
     Connection connection = getConnection();
     Statement statement = connection.createStatement();
